pyproject.toml
[tool.poetry]
name = "project-ia-hematology-scanned-slides-backend"
version = "0.1.0"
description = ""
authors = ["Lucas Praciano <luskcct@gmail.com>"]
readme = "README.md"
documentation = "/docs"
packages = [{ include = "project_ia_hematology_scanned_slides_backend" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
psycopg = { extras = ["binary", "pool"], version = "^3.1.12" }
sqlalchemy = { extras = ["asyncio"], version = "^2.0.23" }
pydantic = "^2.4.2"
dynaconf = "^3.2.4"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.1"
asgi-lifespan = "^2.1.0"
pillow = "^10.1.0"
toml = "^0.10.2"
uvicorn = "^0.24.0"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
ultralytics = "^8.0.207"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
python_files = "tests.py test_*.py *_tests.py"
asyncio_mode = "auto"
filterwarnings = [
    "ignore",
    "ignore:.*U.*mode is deprecated:DeprecationWarning"
]